requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-phonenumber-field==7.1.0
phonenumberslite==8.13.13
psycopg2==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
